+++ greenfield/eldenring/data.py
@@ -5557,4 +5557,4 @@ GESTURE_AWARD_FLAGS = {
     60864: (115, 1075750828, 'O Mother'),
 }
 
-_GEN_STAMP = {'inputs_hash': 'sha256:7063797ea927525e134712162f1df23c5659bc5cb0bfb19501e8201c5a9b97b7', 'module': 'data.py', 'body_sha256': 'sha256:aaf758d0d2de92f871bfc3e6e212c7c75a912bbddafce9595a2a26b55ef326b0'}
+_GEN_STAMP = {'inputs_hash': 'sha256:2c43cab23f441152cd83958ac735642751bf3a538cff1c7c93eb2ace3f0fceb8', 'module': 'data.py', 'body_sha256': 'sha256:aaf758d0d2de92f871bfc3e6e212c7c75a912bbddafce9595a2a26b55ef326b0'}